Move converting loaded settings into dashboards command module
At the end the settings should be completly transparent for the redux
actions.
