+++ ng/src/web/store/dashboard/settings/actions.js
@@ -42,14 +42,6 @@ export const DASHBOARD_SETTINGS_SAVING_ERROR =
 export const DASHBOARD_SETTINGS_SAVING_REQUEST =
   'DASHBOARD_SETTINGS_SAVING_REQUEST';
 
-const settingsV1toDashboardSettings = ({data: rows} = {}) => ({
-  rows: rows.map(({height, data}) =>
-    createRow(data.map(item => createItem({
-      name: item.name,
-      filterId: item.filt_id,
-    })), height)),
-  });
-
 const dashboardSettings2SettingsV1 = ({rows}) => ({
   version: 1,
   data: rows.map(({height, items: rowItems}) => ({
@@ -62,28 +54,6 @@ const dashboardSettings2SettingsV1 = ({rows}) => ({
     })),
   })),
 });
-
-const convertSettings = (settings = {}) => {
-  if (settings.version === 1) {
-    return settingsV1toDashboardSettings(settings);
-  }
-  return settings;
-};
-
-const convertLoadedSettings = (settings = {}) => {
-  /* currently the loaded settings contain an object with settings for all dashboards
-    {
-       dashboardId1: settings1,
-       dashboardId2: settings2,
-    }
-    the format for the settings itself may vary
-  */
-  const converted = {};
-  Object.entries(settings).forEach(([id, value]) => {
-    converted[id] = convertSettings(value);
-  });
-  return converted;
-};
 
 export const receivedDashboardSettings = (id, settings, defaults) => ({
   type: DASHBOARD_SETTINGS_LOADING_SUCCESS,
@@ -133,8 +103,8 @@ export const loadSettings = ({gmp}) => (id, defaults) =>
 
   const promise = gmp.dashboards.currentSettings();
   return promise.then(
-    response => dispatch(receivedDashboardSettings(id,
-      convertLoadedSettings(response.data), defaults)),
+    response => dispatch(receivedDashboardSettings(
+      id, response.data, defaults)),
     error => dispatch(receivedDashboardSettingsLoadingError(error)),
   );
 };
